FEATURE: Automatically add TOC based on topic category / tags (#42)
This PR adds two settings.

One allows admins to select a list of categories that will automatically enable the TOC if the topic category matches.
The other does the same but with tags.

Note that the topic still needs to have the other requirements for TOC, such as having headings. Also, this will still only apply to the first post.

This PR also moves some setting descriptions to the appropriate place.

It also uses `pretender` for the tests instead of generating the fixture manually.