FIX: Scroll enough to show topic timeline from down arrow
This fixes the problem where if you clicked the down / scroll
to bottom arrow at the top of the table of contents, the scroll
would _just_ not quite be enough to show the topic timeline,
which would require another mouse scroll.

Now if there is a second post we just scroll directly to the
top of that, which makes the topic timeline show immediately.